Combine qdq_util and qdq_util_minimal files.
